+++ scenes/metroflip_scene_start.c
@@ -22,6 +22,9 @@ void metroflip_scene_start_on_enter(void* context) {
         app);
 
     submenu_add_item(
+        submenu, "Saved", MetroflipSceneLoad, metroflip_scene_start_submenu_callback, app);
+
+    submenu_add_item(
         submenu,
         "Supported Cards",
         MetroflipSceneSupported,